--- frontend/src/app/(pages)/(account)/login/note.txt
@@ -1,1 +1,0 @@
-TODO: Create User Interface for Login pages -- Delete/Edit after accomplishing.
