+++ frontend/src/app/(pages)/(account)/login/page.tsx
@@ -1,10 +1,14 @@
 import Login from '@/components/pages/AccountGroup/login';
+import Header from '@/components/shared/header';
 
 export default function LoginPage() {
     return (
-        <div className="flex min-h-screen items-center justify-center bg-background px-4">
-            <Login />
-        </div>
+        <>
+            <Header />
+            <div className="flex min-h-screen items-center justify-center bg-background px-4">
+                <Login />
+            </div>
+        </>
     );
 
 };
