+++ frontend/src/app/page.tsx
@@ -1,8 +1,10 @@
 import {LandingPage} from '@/components/pages/LandingPage/landing';
+import Header from '@/components/shared/header';
 
 export default function Landing() {
   return (
     <>
+      <Header />
       <LandingPage />
     </>
   );
