+++ frontend/src/components/pages/AccountGroup/login.tsx
@@ -11,65 +11,74 @@ import {
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { ModeToggle } from "@/components/shared/ModeToggle"
+import Link from "next/link"
 
 export default function LoginCard() {
   return (
     <>
-    <Card className="w-full max-w-sm">
-      <CardHeader>
-        <CardTitle>Log in to your account</CardTitle>
-        <CardDescription>
-          Enter your email below to login to your account
-        </CardDescription>
-        <CardAction>
-            <Button variant="link">
-                <a href="signup">
-                    Sign Up
-                </a>
-            </Button>
-        </CardAction>
-      </CardHeader>
-      <CardContent>
-        <form>
-          <div className="flex flex-col gap-6">
-            <div className="grid gap-2">
-              <Label htmlFor="email">Email</Label>
-              <Input
-                id="email"
-                type="email"
-                placeholder="m@example.com"
-                required
-              />
-            </div>
-            <div className="grid gap-2">
-              <div className="flex items-center">
-                <Label htmlFor="password">Password</Label>
-                <a
-                  href="#"
-                  className="ml-auto inline-block text-sm underline-offset-4 hover:underline"
-                >
-                  Forgot your password?
-                </a>
+      <Card className="w-full max-w-sm">
+        <CardHeader>
+
+          <CardTitle>Log in to your account</CardTitle>
+
+          <CardDescription>
+            Enter your email below to login to your account
+          </CardDescription>
+
+          <CardAction>
+            <Link href="/signup">
+              <Button variant="ghost">Sign Up</Button>
+            </Link>
+          </CardAction>
+
+        </CardHeader>
+        
+        {/* Content */}
+        <CardContent>
+          <form>
+            <div className="flex flex-col gap-6">
+              <div className="grid gap-2">
+                <Label htmlFor="email">Email</Label>
+                <Input
+                  id="email"
+                  type="email"
+                  placeholder="m@example.com"
+                  required
+                />
               </div>
-              <Input id="password" type="password" required />
+              <div className="grid gap-2">
+                <div className="flex items-center">
+                  <Label htmlFor="password">Password</Label>
+                  <a
+                    href="#"
+                    className="ml-auto inline-block text-sm underline-offset-4 hover:underline"
+                  >
+                    Forgot your password?
+                  </a>
+                </div>
+                <Input id="password" type="password" required />
+              </div>
             </div>
-          </div>
-        </form>
-      </CardContent>
-      <CardFooter className="flex-col gap-2">
-        <Button type="submit" className="w-full">
-          Login
-        </Button>
-        <Button variant="outline" className="w-full">
-          Login with Google
-        </Button>
-      </CardFooter>
-    </Card>
+          </form>
+        </CardContent>
 
+        {/* Footer */}
+        <CardFooter className="flex-col gap-2">
+          <Button type="submit" className="w-full">
+            Login
+          </Button>
+          <Button variant="outline" className="w-full">
+            Login with Google
+          </Button>
+        </CardFooter>
 
-    <div className="fixed bottom-4 right-4 text-sm text-muted-foreground">
-      <ModeToggle />
-    </div>
+      </Card>
+
+      {/* Theme Toggle */}
+      <div className="fixed bottom-4 right-4 text-sm text-muted-foreground">
+        <ModeToggle />
+      </div>
+      
     </>
   )
 }
