--- frontend/src/components/pages/LandingPage/header.tsx
@@ -1,14 +1,0 @@
-export default function Header() {
-    return (
-        <header className="w-full px-4 py-2 flex items-center justify-between bg-background dark:bg-background">
-            <div className="text-2xl font-bold text-primary">
-                [Music App Name]
-            </div>
-            <nav className="flex items-center justify-between">
-                <div className="flex items-center space-x-6">
-                    
-                </div>
-            </nav>
-        </header>
-    )
-}
